Split storage replication
