+++ crates/truthdb-core/src/storage/replication/mod.rs
@@ -1,0 +1,7 @@
+mod promotion;
+mod readable_standby;
+mod slots;
+mod streaming;
+mod synchronous_commit;
+
+pub(super) use synchronous_commit::SyncCommitState;
